requirements.txt
django==4.1.2
djangorestframework==3.14.0
django-shortuuidfield==0.1.3
shortuuid~=1.0.9
gunicorn==20.1.0
graphene-django==3.0.0

pip==22.2.2
attrs~=22.1.0
toml~=0.10.2
wheel~=0.37.0
cryptography~=36.0.1
Pillow~=8.4.0
tornado~=6.1
Jinja2~=3.0.3
setuptools~=59.0.1
packaging~=21.3
pyparsing~=3.0.6
future~=0.18.2
graphene~=3.1.1